+++ build.properties
@@ -1,6 +1,6 @@
-mod_version=0.0.0.1
+mod_version=0.0.0.2
 mod_name=TooltipMod
-mod_group=info.mattmc.tooltipmod
+mod_group=com.mattsmeets.tooltip
 mc_version=1.12.1
 forge_version=14.22.0.2462
 mcp_version=snapshot_20171003
