+++ src/main/java/com/mattsmeets/tooltip/CommonProxy.java
@@ -1,0 +1,14 @@
+/*
+  Copyright (c) 2016 / 2017, Matt Smeets and Aroma1997
+  <p>
+  The Tooltipmod is distributed under the terms of the Minecraft Mod Public
+  License 1.0, or MMPL. Please check the contents of the license located in
+  http://www.mod-buildcraft.com/MMPL-1.0.txt
+ */
+package com.mattsmeets.tooltip;
+
+public class CommonProxy {
+
+	public void init() {}
+
+}
