+++ src/main/java/com/mattsmeets/tooltip/Reference.java
@@ -1,11 +1,11 @@
-/**
- * Copyright (c) 2016, MattsOnMc and Aroma1997
- * <p>
- * The Tooltipmod is distributed under the terms of the Minecraft Mod Public
- * License 1.0, or MMPL. Please check the contents of the license located in
- * http://www.mod-buildcraft.com/MMPL-1.0.txt
+/*
+  Copyright (c) 2016 / 2017, Matt Smeets and Aroma1997
+  <p>
+  The Tooltipmod is distributed under the terms of the Minecraft Mod Public
+  License 1.0, or MMPL. Please check the contents of the license located in
+  http://www.mod-buildcraft.com/MMPL-1.0.txt
  */
-package info.mattmc.tooltipmod;
+package com.mattsmeets.tooltip;
 
 import java.io.InputStream;
 import java.util.Properties;
