+++ src/main/java/com/mattsmeets/tooltip/TooltipHelper.java
@@ -1,12 +1,11 @@
-/**
- * Copyright (c) 2016, MattsOnMc and Aroma1997
- * <p>
- * The Tooltipmod is distributed under the terms of the Minecraft Mod Public
- * License 1.0, or MMPL. Please check the contents of the license located in
- * http://www.mod-buildcraft.com/MMPL-1.0.txt
+/*
+  Copyright (c) 2016 / 2017, Matt Smeets and Aroma1997
+  <p>
+  The Tooltipmod is distributed under the terms of the Minecraft Mod Public
+  License 1.0, or MMPL. Please check the contents of the license located in
+  http://www.mod-buildcraft.com/MMPL-1.0.txt
  */
-
-package info.mattmc.tooltipmod;
+package com.mattsmeets.tooltip;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,7 +16,6 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTBase;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.nbt.NBTTagList;
-
 import net.minecraftforge.fml.common.registry.ForgeRegistries;
 import net.minecraftforge.oredict.OreDictionary;
