+++ src/main/java/com/mattsmeets/tooltip/TooltipMod.java
@@ -1,13 +1,13 @@
-/**
- * Copyright (c) 2016, MattsOnMc and Aroma1997
- * <p>
- * The Tooltipmod is distributed under the terms of the Minecraft Mod Public
- * License 1.0, or MMPL. Please check the contents of the license located in
- * http://www.mod-buildcraft.com/MMPL-1.0.txt
+/*
+  Copyright (c) 2016 / 2017, Matt Smeets and Aroma1997
+  <p>
+  The Tooltipmod is distributed under the terms of the Minecraft Mod Public
+  License 1.0, or MMPL. Please check the contents of the license located in
+  http://www.mod-buildcraft.com/MMPL-1.0.txt
  */
-package info.mattmc.tooltipmod;
+package com.mattsmeets.tooltip;
 
-import info.mattmc.tooltipmod.commands.CommandTooltip;
+import com.mattsmeets.tooltip.commands.CommandTooltip;
 
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
@@ -17,14 +17,13 @@ import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLServerStartingEvent;
 
-
 @Mod(modid = Reference.MOD_ID, name = Reference.MOD_NAME)
 public class TooltipMod {
 
 	@Instance
 	public static TooltipMod instance;
 
-	@SidedProxy(clientSide = "info.mattmc.tooltipmod.client.ClientProxy", serverSide = "info.mattmc.tooltipmod.CommonProxy")
+	@SidedProxy(clientSide = "com.mattsmeets.tooltip.client.ClientProxy", serverSide = "com.mattsmeets.tooltip.CommonProxy")
 	public static CommonProxy proxy;
 
 	@EventHandler
