+++ src/main/java/com/mattsmeets/tooltip/commands/CommandBlockInfo.java
@@ -1,8 +1,17 @@
-package info.mattmc.tooltipmod.commands;
+/*
+  Copyright (c) 2016 / 2017, Matt Smeets and Aroma1997
+  <p>
+  The Tooltipmod is distributed under the terms of the Minecraft Mod Public
+  License 1.0, or MMPL. Please check the contents of the license located in
+  http://www.mod-buildcraft.com/MMPL-1.0.txt
+ */
+package com.mattsmeets.tooltip.commands;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import aroma1997.core.command.AromaBaseCommand;
+import aroma1997.core.util.ServerUtil;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommandSender;
@@ -12,9 +21,6 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
-
-import aroma1997.core.command.AromaBaseCommand;
-import aroma1997.core.util.ServerUtil;
 
 public class CommandBlockInfo extends AromaBaseCommand {
 	@Override
