+++ src/main/java/com/mattsmeets/tooltip/commands/CommandItemInfo.java
@@ -1,16 +1,21 @@
-package info.mattmc.tooltipmod.commands;
+/*
+  Copyright (c) 2016 / 2017, Matt Smeets and Aroma1997
+  <p>
+  The Tooltipmod is distributed under the terms of the Minecraft Mod Public
+  License 1.0, or MMPL. Please check the contents of the license located in
+  http://www.mod-buildcraft.com/MMPL-1.0.txt
+ */
+package com.mattsmeets.tooltip.commands;
 
-
-import info.mattmc.tooltipmod.TooltipHelper;
-
+import aroma1997.core.command.AromaBaseCommand;
+import aroma1997.core.util.ServerUtil;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.MinecraftServer;
 
-import aroma1997.core.command.AromaBaseCommand;
-import aroma1997.core.util.ServerUtil;
+import com.mattsmeets.tooltip.TooltipHelper;
 
 public class CommandItemInfo extends AromaBaseCommand {
