+++ src/main/java/com/mattsmeets/tooltip/commands/CommandTileEntityFieldInfo.java
@@ -1,15 +1,21 @@
-package info.mattmc.tooltipmod.commands;
+/*
+  Copyright (c) 2016 / 2017, Matt Smeets and Aroma1997
+  <p>
+  The Tooltipmod is distributed under the terms of the Minecraft Mod Public
+  License 1.0, or MMPL. Please check the contents of the license located in
+  http://www.mod-buildcraft.com/MMPL-1.0.txt
+ */
+package com.mattsmeets.tooltip.commands;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 import java.util.List;
 
+import aroma1997.core.coremod.asm.AsmUtils;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-
-import aroma1997.core.coremod.asm.AsmUtils;
 
 public class CommandTileEntityFieldInfo extends CommandBlockInfo {
