+++ src/main/java/com/mattsmeets/tooltip/commands/CommandTileEntityNBTInfo.java
@@ -1,9 +1,16 @@
-package info.mattmc.tooltipmod.commands;
+/*
+  Copyright (c) 2016 / 2017, Matt Smeets and Aroma1997
+  <p>
+  The Tooltipmod is distributed under the terms of the Minecraft Mod Public
+  License 1.0, or MMPL. Please check the contents of the license located in
+  http://www.mod-buildcraft.com/MMPL-1.0.txt
+ */
+package com.mattsmeets.tooltip.commands;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import info.mattmc.tooltipmod.TooltipHelper;
+import com.mattsmeets.tooltip.TooltipHelper;
 
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
