--- src/main/java/info/mattmc/tooltipmod/CommonProxy.java
@@ -1,14 +1,0 @@
-/**
- * Copyright (c) 2016, MattsOnMc and Aroma1997
- * <p>
- * The Tooltipmod is distributed under the terms of the Minecraft Mod Public
- * License 1.0, or MMPL. Please check the contents of the license located in
- * http://www.mod-buildcraft.com/MMPL-1.0.txt
- */
-package info.mattmc.tooltipmod;
-
-public class CommonProxy {
-
-	public void init() {}
-
-}
